ci: move cache handling for state image to master branch only
Before this commit, we split cache handling between the pull request
branch and the master branch, which complicates things and allows
bleedover between the branches.

After this commit, only deal with cache on master branch.
This simplifies the cache operations.

During CI build, we build the images and tag them with ":ci".
During publish, we compare the "vcs-ref" labels of the ci and cached
state images. If the labels differ, then we:

1. re-label the ci state image
2. push it
3. destroy the cache
4. save the cache